backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
databases
python-dotenv
openai
aiohttp
python-multipart
google-generativeai